requirements.txt
faiss-cpu==1.9.0.post1
Flask==3.1.0
sentence-transformers==3.3.1